# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API responsável por gerar prompts e fazer upload no bucket.",
        "title": "Generate Prompt Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte Syphertech",
            "email": "suporte@syphertech.com.br"
        },
        "version": "1.0"
    },
    "host": "generate-prompt.syphertech.com.br",
    "basePath": "/",
    "paths": {
        "/api/v1/agent-number": {
            "post": {
                "description": "Insere um novo mapeamento client_id e número de WhatsApp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-number"
                ],
                "summary": "Criar AgentNumber",
                "parameters": [
                    {
                        "description": "Dados do agente",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/agentnumber.AgentNumber"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Criado com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Payload inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/generate-prompt": {
            "post": {
                "description": "Recebe os dados de prompt e realiza o upload em bucket S3 com metadados automáticos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "Geração de Prompt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do bucket",
                        "name": "bucket",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Dados do Prompt",
                        "name": "prompt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Prompt"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Prompt gerado com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida ou campos obrigatórios ausentes",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Método HTTP não permitido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao processar prompt",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "agentnumber.AgentNumber": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "whatsapp_number": {
                    "type": "string"
                }
            }
        },
        "domain.Metadados": {
            "type": "object",
            "properties": {
                "dataCriacao": {
                    "type": "string"
                },
                "origem": {
                    "type": "string"
                },
                "ultimaAtualizacao": {
                    "type": "string"
                }
            }
        },
        "domain.Prompt": {
            "type": "object",
            "properties": {
                "clienteID": {
                    "type": "string"
                },
                "idiomas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "metadados": {
                    "$ref": "#/definitions/domain.Metadados"
                },
                "modeloIA": {
                    "type": "string"
                },
                "nomeEmpresa": {
                    "type": "string"
                },
                "servicos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Servico"
                    }
                },
                "systemPrompt": {
                    "type": "string"
                },
                "versaoPrompt": {
                    "type": "integer"
                }
            }
        },
        "domain.Servico": {
            "type": "object",
            "properties": {
                "descricao": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "preco": {
                    "type": "number"
                }
            }
        }
    }
}